feat(todo): add button to generate AI-powered subtasks
This commit introduces a new feature that allows users to generate AI-powered subtasks for a todo item using a button. The `onGenerateSubtasks` function is added to the `TodoListProps` interface and implemented in the `Todo` component. A button with a sparkle icon is added to the `TodoList` component to trigger the generation of subtasks. This enhances productivity by automating the creation of subtasks based on the todo's text.